+++ src/blameandshame/util.py
@@ -1,17 +1,6 @@
-from enum import Enum
 import git
+from blameandshame.base import Change
 from typing import FrozenSet, List, Tuple, Optional, Set
-
-
-class Change(Enum):
-    """
-    Enum of the possible types of git changes. These values can be used as
-    arguments to the Diff object's iter_change_type method.
-    """
-    ADDED = 'A'
-    DELETED = 'D'
-    MODIFIED = 'M'
-    RENAMED = 'R'
 
 
 def files_in_commit(repo: git.Repo,
